requirements.txt
absl-py==0.13.0
addict==2.4.0
antlr4-python3-runtime==4.8
anyio==3.3.4
argon2-cffi==21.1.0
atari-py==0.2.9
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.2
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.4
colorlog==6.5.0
commonmark==0.9.1
configparser==5.0.2
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
deprecation==2.1.0
docker-pycreds==0.4.0
einops==0.3.0
entrypoints==0.3
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.33.0
google-auth-oauthlib==0.4.4
grpcio==1.38.1
gym==0.18.3
hydra-core==1.1.1
idna==3.2
igl @ git+https://github.com/libigl/libigl-python-bindings@3c3a6ea8f6c95720f76c9b4d8c5a26b6c8330235
imageio==2.13.3
importlib-resources==5.2.2
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
json5==0.9.6
jsonschema==4.1.2
jupyter-client==7.0.6
jupyter-core==4.9.0
jupyter-packaging==0.11.0
jupyter-server==1.11.1
jupyterlab==3.2.1
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.3.3
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.5
numpy==1.21.3
oauthlib==3.1.1
omegaconf==2.1.1
open3d==0.13.0
opencv-python==4.5.3.56
packaging==21.0
pandas==1.3.0
pandocfilters==1.5.0
parso==0.8.2
pathtools==0.1.2
perceiver-pytorch==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.21
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyglet==1.5.15
Pygments==2.10.0
pyparsing==2.4.7
PyQt6==6.2.1
PyQt6-Qt6==6.2.1
PyQt6-sip==13.1.0
pyqtgraph==0.12.3
pyrsistent==0.18.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rich==10.12.0
robust-laplacian==0.2.2
rsa==4.7.2
scikit-learn==1.0.1
scipy==1.7.0
Send2Trash==1.8.0
sentry-sdk==1.3.0
shortuuid==1.0.1
six==1.16.0
smmap==4.0.0
sniffio==1.2.0
stable-baselines3==1.1.0
subprocess32==3.5.4
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tomlkit==0.7.2
torch==1.9.0
tornado==6.1
tqdm==4.62.2
traitlets==5.1.1
typing-extensions==3.10.0.0
urllib3==1.26.6
wandb==0.11.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
zipp==3.5.0
